End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,12 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
   selector: 'app-messages',
   templateUrl: './messages.component.html',
-  styleUrls: ['./messages.component.css']
+  styleUrls: ['./messages.component.css'],
 })
 export class MessagesComponent implements OnInit {
   messages: Message[] = [];
@@ -16,29 +17,46 @@ export class MessagesComponent implements OnInit {
   container = 'Unread';
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService
+  ) {}
 
   ngOnInit(): void {
     this.loadMessages();
   }
 
-  loadMessages(): void{
+  loadMessages(): void {
     this.loading = true;
-    this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(response => {
-      this.messages = response.result;
-      this.pagination = response.pagination;
-      this.loading = false;
-    });
+    this.messageService
+      .getMessages(this.pageNumber, this.pageSize, this.container)
+      .subscribe((response) => {
+        this.messages = response.result;
+        this.pagination = response.pagination;
+        this.loading = false;
+      });
   }
 
   deleteMessage(id: number): void {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-      this.loadMessages();
-    });
+    this.confirmService
+      .confirm(
+        'Confirm Delete',
+        'Are you sure you want to delete? This cannot be undone.'
+      )
+      .subscribe((result) => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(
+              this.messages.findIndex((m) => m.id === id),
+              1
+            );
+            this.loadMessages();
+          });
+        }
+      });
   }
 
-  pageChanged(event: any): void{
+  pageChanged(event: any): void {
     this.pageNumber = event.page;
     this.loadMessages();
   }
